Liqun/plugin only (#91)
Add the plugin-only mode which is now only supported by openai models.
This is now configured by setting `session.plugin_only_mode=True`. I
have removed the old configurations and verifications for the
plugin-only mode in the code interpreter (and generator). Now, a new
code interpreter and generator are created to handle this mode which
relies on the function calling feature of the LLMs.